+++ app/Listeners/PharmacySellingListener.php
@@ -35,6 +35,10 @@ class PharmacySellingListener
             return true;
         }
 
+        if (!in_array($sale->getAttribute('state'),[Sale::INITIATED])) {
+            return true;
+        }
+
         if ($event::TYPE == Sale::REPROCESSED){
             ($sale->dues())?$sale->dues()->delete():null;
             ($sale->accredits())?$sale->accredits()->delete():null;
